+++ lib/oas_rails.rb
@@ -45,7 +45,11 @@ module OasRails
   end
 
   module YARD
-    autoload :OasYARDFactory, 'oas_rails/yard/oas_yard_factory'
+    autoload :RequestBodyTag, 'oas_rails/yard/request_body_tag'
+    autoload :RequestBodyExampleTag, 'oas_rails/yard/request_body_example_tag'
+    autoload :ParameterTag, 'oas_rails/yard/parameter_tag'
+    autoload :ResponseTag, 'oas_rails/yard/response_tag'
+    autoload :OasRailsFactory, 'oas_rails/yard/oas_rails_factory'
   end
 
   module Extractors
@@ -73,7 +77,7 @@ module OasRails
     end
 
     def configure_yard!
-      ::YARD::Tags::Library.default_factory = YARD::OasYARDFactory
+      ::YARD::Tags::Library.default_factory = YARD::OasRailsFactory
       yard_tags = {
         'Request body' => [:request_body, :with_request_body],
         'Request body Example' => [:request_body_example, :with_request_body_example],
