+++ lib/oas_rails/yard/request_body_example_tag.rb
@@ -1,0 +1,12 @@
+module OasRails
+  module YARD
+    class RequestBodyExampleTag < ::YARD::Tags::Tag
+      attr_accessor :content
+
+      def initialize(tag_name, text, content: {})
+        super(tag_name, text, nil, nil)
+        @content = content
+      end
+    end
+  end
+end
